rename services closer to our pattern
