+++ packages/backend/src/nest/auth/services/dm/dm.service.ts
@@ -3,20 +3,19 @@
  */
 
 import { createHash } from 'crypto'
-import { BaseChainService } from '../baseChainService'
+import { ChainServiceBase } from '../chainServiceBase'
 import { Keyset } from '@localfirst/auth'
 import { SigChain } from '../../sigchain'
 import { createLogger } from '../../../common/logger'
 
 const logger = createLogger('auth:dmService')
-class DMService extends BaseChainService {
+class DMService extends ChainServiceBase {
   private dmMap: Map<string, string[]> = new Map()
 
   public static init(sigChain: SigChain): DMService {
     return new DMService(sigChain)
   }
 
-  // TODO: incorporate persistence
   public create(memberIds: string[]): string {
     // this is the ID that will be used for storage and mapping purposes
     const id = DMService.getDmId(memberIds)
