+++ packages/backend/src/nest/auth/sigchain.ts
@@ -4,13 +4,13 @@
 
 import * as auth from '@localfirst/auth'
 import { LoadedSigChain } from './types'
-import { UserService } from './services/members/userService'
-import { RoleService } from './services/roles/roleService'
-import { ChannelService } from './services/roles/channelService'
-import { DeviceService } from './services/members/deviceService'
-import { InviteService } from './services/invites/inviteService'
-import { DMService } from './services/dm/dmService'
-import { CryptoService } from './services/crypto/cryptoService'
+import { UserService } from './services/members/user.service'
+import { RoleService } from './services/roles/role.service'
+import { ChannelService } from './services/roles/channel.service'
+import { DeviceService } from './services/members/device.service'
+import { InviteService } from './services/invites/invite.service'
+import { DMService } from './services/dm/dm.service'
+import { CryptoService } from './services/crypto/crypto.service'
 import { RoleName } from './services/roles/roles'
 import { createLogger } from '../common/logger'
